+++ src/main/java/org/apiphany/client/http/AbstractHttpExchangeClient.java
@@ -6,6 +6,7 @@ import java.util.Map;
 
 import javax.net.ssl.SSLContext;
 
+import org.apiphany.ApiMimeType;
 import org.apiphany.ApiRequest;
 import org.apiphany.client.ClientProperties;
 import org.apiphany.client.ContentConverter;
@@ -13,7 +14,6 @@ import org.apiphany.header.HeaderValuesChain;
 import org.apiphany.header.Headers;
 import org.apiphany.header.MapHeaderValues;
 import org.apiphany.http.HttpHeaderValues;
-import org.apiphany.http.ResolvedContentType;
 import org.apiphany.http.TracingHeader;
 import org.apiphany.json.JsonBuilder;
 import org.apiphany.json.jackson.JacksonJsonHttpContentConverter;
@@ -118,13 +118,13 @@ public abstract class AbstractHttpExchangeClient implements HttpExchangeClient {
 	 * @param <H> the type of response headers
 	 *
 	 * @param apiRequest the API request containing response type information
-	 * @param contentType the resolved content type
+	 * @param contentType the HTTP content type
 	 * @param headers the response headers used for content type negotiation
 	 * @param body the raw response body to be converted
 	 * @return the converted body of type {@code U}
 	 * @throws UnsupportedOperationException if no compatible content converter is found
 	 */
-	protected <T, U, H> U convertBody(final ApiRequest<T> apiRequest, final ResolvedContentType contentType, final H headers, final Object body) {
+	protected <T, U, H> U convertBody(final ApiRequest<T> apiRequest, final ApiMimeType contentType, final H headers, final Object body) {
 		if (body instanceof byte[] bytes && bytes.length == 0) {
 			return null;
 		}