+++ src/main/java/org/apiphany/json/jackson/JacksonJsonHttpContentConverter.java
@@ -1,11 +1,11 @@
 package org.apiphany.json.jackson;
 
 import org.apiphany.ApiMessage;
+import org.apiphany.ApiMimeType;
 import org.apiphany.client.ContentConverter;
 import org.apiphany.client.http.HttpContentConverter;
 import org.apiphany.header.HeaderValuesChain;
-import org.apiphany.http.ContentType;
-import org.apiphany.http.ResolvedContentType;
+import org.apiphany.io.ContentType;
 import org.morphix.reflection.GenericClass;
 
 /**
@@ -30,19 +30,14 @@ public class JacksonJsonHttpContentConverter<T> implements HttpContentConverter<
 	 * Converts the given object to an instance of the specified class. This method supports conversion from JSON strings.
 	 *
 	 * @param obj the object to convert
+	 * @param mimeType the mime type
 	 * @param dstClass the target class to which the JSON content will be de-serialized
 	 * @return the de-serialized object of type {@code T}
 	 * @throws UnsupportedOperationException if the input object is not a JSON string
 	 */
 	@Override
-	public T from(final Object obj, final ResolvedContentType contentType, final Class<T> dstClass) {
-		if (obj instanceof String str) {
-			return JacksonJsonBuilder.fromJson(str, dstClass);
-		}
-		if (obj instanceof byte[] bytes) {
-			return JacksonJsonBuilder.fromJson(bytes, dstClass);
-		}
-		throw new UnsupportedOperationException("Conversion from " + obj.getClass().getCanonicalName() + " is not supported.");
+	public T from(final Object obj, final ApiMimeType mimeType, final Class<T> dstClass) {
+		return JacksonJsonBuilder.fromJson(obj, dstClass);
 	}
 
 	/**
@@ -52,17 +47,11 @@ public class JacksonJsonHttpContentConverter<T> implements HttpContentConverter<
 	 * @param obj the object to convert, which must be a JSON string
 	 * @param genericDstClass the target generic class to which the JSON content will be de-serialized
 	 * @return the de-serialized object of type {@code T}
-	 * @throws UnsupportedOperationException if the input object is not a JSON string
+	 * @throws UnsupportedOperationException if the input object is not a JSON
 	 */
 	@Override
-	public T from(final Object obj, final ResolvedContentType contentType, final GenericClass<T> genericDstClass) {
-		if (obj instanceof String str) {
-			return JacksonJsonBuilder.fromJson(str, genericDstClass);
-		}
-		if (obj instanceof byte[] bytes) {
-			return JacksonJsonBuilder.fromJson(bytes, genericDstClass);
-		}
-		throw new UnsupportedOperationException("Conversion from " + obj.getClass().getCanonicalName() + " is not supported.");
+	public T from(final Object obj, final ApiMimeType mimeType, final GenericClass<T> genericDstClass) {
+		return JacksonJsonBuilder.fromJson(obj, genericDstClass);
 	}
 
 	/**
@@ -73,15 +62,15 @@ public class JacksonJsonHttpContentConverter<T> implements HttpContentConverter<
 	 * @param <V> the type of the headers
 	 *
 	 * @param message the {@link ApiMessage} containing the content to convert
-	 * @param resolvedContentTyoe the resolved content type
+	 * @param mimeType the content type
 	 * @param headers the headers that may influence the conversion
 	 * @return true if the content type is {@code application/json}, false otherwise
 	 */
 	@Override
-	public <U, V> boolean isConvertible(final ApiMessage<U> message, final ResolvedContentType resolvedContentTyoe, final V headers,
+	public <U, V> boolean isConvertible(final ApiMessage<U> message, final ApiMimeType mimeType, final V headers,
 			final HeaderValuesChain headerValuesChain) {
-		if (null != resolvedContentTyoe) {
-			return ContentType.APPLICATION_JSON == resolvedContentTyoe.contentType();
+		if (null != mimeType) {
+			return ContentType.APPLICATION_JSON == mimeType.contentType();
 		}
 		return false;
 	}