+++ src/main/java/org/apiphany/security/AuthenticationException.java
@@ -19,7 +19,7 @@ public class AuthenticationException extends HttpException {
 	private static final long serialVersionUID = 591403103621076702L;
 
 	/**
-	 * Constructor with an exception message.
+	 * Constructor with an exception message. The HTTP status will default to {@link HttpStatus#UNAUTHORIZED}.
 	 *
 	 * @param message the exception message
 	 */
@@ -28,7 +28,16 @@ public class AuthenticationException extends HttpException {
 	}
 
 	/**
-	 * Constructor with exception message and cause.
+	 * Constructor with an exception message. The HTTP status will default to {@link HttpStatus#UNAUTHORIZED}.
+	 *
+	 * @param cause the cause of the exception
+	 */
+	public AuthenticationException(final Throwable cause) {
+		super(HttpStatus.UNAUTHORIZED, null, cause);
+	}
+
+	/**
+	 * Constructor with exception message and cause. The HTTP status will default to {@link HttpStatus#UNAUTHORIZED}.
 	 *
 	 * @param message the exception message
 	 * @param cause cause of the exception